+++ src/views/LearnType/Hindi-Inscript/Learn-Inscript-Hindi.vue
@@ -225,7 +225,7 @@ export default {
 /** Typing Content Div */
 #thisClass {
   width: 96.4%;
-  height: 113px;
+  height: 200px;
   overflow-y: scroll;
 }